pyproject.toml
[tool.poetry]
name = "ipwndfu"
version = "2.0.0b3"
description = "The DFU exploitation toolkit for Apple devices"
readme = "README.md"

packages = [
    { include = "ipwndfu", from="src" },
    { include = "libusbfinder", from="src" }
]

authors = ["axi0mX <axi0mXor@gmail.com>"]

[tool.isort]
profile = "black"

[tool.mypy]
strict = true
exclude = "tests"

[tool.mypy-usb]
ignore_missing_imports = true

[tool.poetry.scripts]
ipwndfu = "ipwndfu.main:main"
ibootpatcher = "ipwndfu.ibootpatcher:main"

[tool.poetry.dependencies]
python = "^3.8"
cryptography = "^36.0.1"
pyusb = "^1.2.1"

[tool.poetry.dev-dependencies]
pylint = "^2.12.2"
pytest = "^7.0.1"
mypy = "^0.931"
pre-commit = "^2.17.0"
pytest-cov = "^3.0.0"
pytest-mypy = "^0.9.1"
black = "^22.1.0"
flake8 = "^4.0.1"
flake8-broken-line = "^0.4.0"
flake8-bugbear = "^22.1.11"
flake8-comprehensions = "^3.8.0"
flake8-eradicate = "^1.2.0"
flake8-no-pep420 = "^2.2.0"
flake8-quotes = "^3.3.1"
flake8-simplify = "^0.17.1"
flake8-tidy-imports = "^4.6.0"
flake8-typing-imports = "^1.12.0"
flake8-use-fstring = "^1.3"
pep8-naming = "^0.12.1"
types-cryptography = "^3.3.15"
MonkeyType = "^21.5.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
